exclude get_coords func from main, tweak it transfer get_coords to utilities/, make function to return one address to lighten the response, separate query with commas to make queries more precise (according to documentation)
